refactor: split generation output writer
